mindpic/requirements.txt
keyboard==0.13.5
pillow==12.0.0
pystray==0.19.5
